Move DefaultMapMarkerIcon to index file and add UserLocationIcon comment
